修改hton和ntoh部分，和修bug

+++ tests/test_network.cpp
@@ -35,13 +35,13 @@ TEST(TestNetwork, ntoh_hton) {
 
 TEST(TestNetwork, vector_append) {
     data_type data1;
-    vector_mem_order_append(data1, 0x01030405u,
+    vector_mem_order_append(data1, static_cast<std::uint32_t>(0x01030405u),
                             static_cast<std::uint16_t>(0x1517u),
                             array_data_type<5>{0x01u, 0x02u, 0x05u, 0x09u, 0x11u},
                             data_type{0x10u, 0x21u, 0x34u, 0x57u},
                             static_cast<std::uint16_t>(0x0115u),
                             static_cast<std::uint8_t>(0x99u),
-                            0x9998979695949392llu);
+                            static_cast<std::uint64_t>(0x9998979695949392llu));
     if constexpr (is_little_endian()) {
         ASSERT_TRUE((
             data1==data_type{0x05u,0x04u,0x03u,0x01u,0x17u,0x15u,0x01u, 0x02u, 0x05u, 0x09u, 0x11u,0x10u, 0x21u, 0x34u,
@@ -57,13 +57,13 @@ TEST(TestNetwork, vector_append) {
 
 
     data_type data2;
-    vector_append_to_net(data2, 0x01030405u,
+    vector_append_to_net(data2, static_cast<std::uint32_t>(0x01030405u),
                          static_cast<std::uint16_t>(0x1517u),
                          array_data_type<5>{0x01u, 0x02u, 0x05u, 0x09u, 0x11u},
                          data_type{0x10u, 0x21u, 0x34u, 0x57u},
                          static_cast<std::uint16_t>(0x0115u),
                          static_cast<std::uint8_t>(0x99u),
-                         0x9998979695949392llu);
+                         static_cast<std::uint64_t>(0x9998979695949392llu));
     ASSERT_TRUE((
         data2==data_type{ 0x01u, 0x03u, 0x04u, 0x05u, 0x15u, 0x17u, 0x01u, 0x02u, 0x05u, 0x09u, 0x11u,0x10u, 0x21u,
         0x34u, 0x57u, 0x01u, 0x15u,0x99u,0x99u,0x98u,0x97u,0x96u,0x95u,0x94u,0x93u,0x92u }
@@ -71,24 +71,24 @@ TEST(TestNetwork, vector_append) {
 }
 
 TEST(TestNetwork, to_network) {
-    auto array = to_network_array(0x01030405u,
+    auto array = to_network_array(static_cast<std::uint32_t>(0x01030405u),
                                   static_cast<std::uint16_t>(0x1517),
                                   array_data_type<5>{0x01u, 0x02u, 0x05u, 0x09u, 0x11u},
                                   static_cast<std::uint16_t>(0x0115u),
                                   static_cast<std::uint8_t>(0x99u),
-                                  0x9998979695949392llu);
+                                  static_cast<std::uint64_t>(0x9998979695949392llu));
     ASSERT_TRUE((
         array==decltype(array){ 0x01u, 0x03u, 0x04u, 0x05u, 0x15u, 0x17u, 0x01u, 0x02u, 0x05u, 0x09u, 0x11u, 0x01u,
         0x15u,0x99u,0x99u,0x98u,0x97u,0x96u,0x95u,0x94u,0x93u,0x92u }
     ));
 
-    auto vec = to_network_data(0x01030405u,
+    auto vec = to_network_data(static_cast<std::uint32_t>(0x01030405u),
                                static_cast<std::uint16_t>(0x1517u),
                                array_data_type<5>{0x01u, 0x02u, 0x05u, 0x09u, 0x11u},
                                data_type{0x10u, 0x21u, 0x34u, 0x57u},
                                static_cast<std::uint16_t>(0x0115u),
                                static_cast<std::uint8_t>(0x99u),
-                               0x9998979695949392llu);
+                               static_cast<std::uint64_t>(0x9998979695949392llu));
     ASSERT_TRUE((
         vec==data_type{ 0x01u, 0x03u, 0x04u, 0x05u, 0x15u, 0x17u, 0x01u, 0x02u, 0x05u, 0x09u, 0x11u,0x10u, 0x21u,
         0x34u, 0x57u, 0x01u, 0x15u,0x99u,0x99u,0x98u,0x97u,0x96u,0x95u,0x94u,0x93u,0x92u }
